[Refector] DropdownItems 삭제
