+++ src/components/pages/Home.tsx
@@ -4,8 +4,9 @@ import { LayoutTemplate } from "../templates"
 export const Home = () => {
     return (
         <>
-            <LayoutTemplate></LayoutTemplate>
-            <MessageList></MessageList>
+            <LayoutTemplate>
+                <MessageList></MessageList>
+            </LayoutTemplate>
         </>
     )
 }
